Document express vs http routes
I'm trying a new thing, where I wrote the end of the literate file as
though it were a blog post that indicated there would be more writing in
the series. This sets me up for the flow wherein I go to a particular
commit, export it as code + docs, and ta da it's a tutorial to publish.
Seeing as you can write, you should really be turning some of this "I'm
copying shit and annotating it" into public documentation. To the commons!